+++ metatensor-torch/CHANGELOG.md
@@ -24,6 +24,7 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 - `read_model_metadata` to load only the `ModelMetadata` from an exported
   atomistic model without having to load the whole model.
 - Users can now store arbitrary additional metadata in `ModelMetadata.extra`
+- Added `Labels.select` function to sub-select entries in labels
 
 ## [Version 0.5.3](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-torch-v0.5.3) - 2024-07-15